Added description migration and fixed bug

+++ api/ExpressedRealms.DB/Models/Skills/SkillLevelDescriptionMapping.cs
@@ -2,8 +2,8 @@ namespace ExpressedRealms.DB.Models.Skills;
 
 public class SkillLevelDescriptionMapping
 {
-    public int SkillLevelId { get; set; }
-    public int SkillTypeId { get; set; }
+    public byte SkillLevelId { get; set; }
+    public byte SkillTypeId { get; set; }
     public string Description { get; set; }
     
     public virtual SkillLevel SkillLevel { get; set; } = null!;
